+++ src/main/java/com/example/voebb/controller/web/AboutPageController.java
@@ -1,0 +1,13 @@
+package com.example.voebb.controller.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AboutPageController {
+
+    @GetMapping("/about")
+    public String getAboutPage() {
+        return "pages/about-us";
+    }
+}
