+++ src/main/resources/templates/admin/fragments/_item-create.html
@@ -188,6 +188,6 @@
         </div>
     </div>
 
-    <script th:src="@{/js/create-item-modal.js}"></script>
+    <script th:src="@{/static/create-item-modal.js}"></script>
 
 </div>
